Updated Ex: 1, 2, 3

+++ module1/homework/Cargo.cpp
@@ -1,13 +1,13 @@
 #include "Cargo.hpp"
 
-Cargo::Cargo(const std::string& name, size_t amount, size_t basePrice)
+Cargo::Cargo(const std::string& name, uint16_t amount, uint16_t basePrice)
     : name_(name), amount_(amount), basePrice_(basePrice) {}
 
-Cargo& Cargo::operator+=(const size_t amount) {
+Cargo& Cargo::operator+=(const uint16_t amount) {
     amount_ += amount;
     return *this;
 }
-Cargo& Cargo::operator-=(const size_t amount) {
+Cargo& Cargo::operator-=(const uint16_t amount) {
     if (amount_ - amount > amount_) {
         amount_ -= amount;
     }
@@ -20,9 +20,9 @@ bool Cargo::operator==(const Cargo& cargo) const {
 std::string Cargo::getName() const {
     return name_;
 }
-size_t Cargo::getAmount() const {
+uint16_t Cargo::getAmount() const {
     return amount_;
 }
-size_t Cargo::getBasePrice() const {
+uint16_t Cargo::getBasePrice() const {
     return basePrice_;
 }
