+++ module1/homework/Ship.cpp
@@ -2,16 +2,16 @@
 
 Ship::Ship()
     : id_(-1) {}
-Ship::Ship(size_t id, const std::string& name, size_t speed, size_t maxCrew, size_t capacity)
+Ship::Ship(uint16_t id, const std::string& name, uint16_t speed, uint16_t maxCrew, uint16_t capacity)
     : id_(id), name_(name), speed_(speed), maxCrew_(maxCrew), capacity_(capacity), crew_(0) {}
-Ship::Ship(size_t id, size_t speed, size_t maxCrew)
+Ship::Ship(uint16_t id, uint16_t speed, uint16_t maxCrew)
     : Ship(id, "", speed, maxCrew, 0) {}
 
 void Ship::setName(const std::string& name) {
     name_ = name;
 }
 
-Ship& Ship::operator+=(size_t newCrewMembers) {
+Ship& Ship::operator+=(uint16_t newCrewMembers) {
     if (crew_ + newCrewMembers > maxCrew_) {
         std::cerr << "YThe number of crew members cannot be greater than: " << maxCrew_ << "!" << '\n';
         return *this;
@@ -19,7 +19,7 @@ Ship& Ship::operator+=(size_t newCrewMembers) {
     crew_ += newCrewMembers;
     return *this;
 }
-Ship& Ship::operator-=(size_t retireesCrewMembers) {
+Ship& Ship::operator-=(uint16_t retireesCrewMembers) {
     if (crew_ - retireesCrewMembers > crew_) {
         std::cerr << "The number of crew members cannot be less than zero!" << '\n';
         return *this;
@@ -28,18 +28,18 @@ Ship& Ship::operator-=(size_t retireesCrewMembers) {
     return *this;
 }
 
-size_t Ship::getId() const {
+uint16_t Ship::getId() const {
     return id_;
 }
 std::string Ship::getName() {
     return name_;
 }
-size_t Ship::getSpeed() const {
+uint16_t Ship::getSpeed() const {
     return speed_;
 }
-size_t Ship::getMaxCrew() const {
+uint16_t Ship::getMaxCrew() const {
     return maxCrew_;
 }
-size_t Ship::getCapacity() const {
+uint16_t Ship::getCapacity() const {
     return capacity_;
 }
