+++ module1/homework/Ship.hpp
@@ -4,26 +4,26 @@
 #include <string>
 
 class Ship {
-    size_t id_;
+    uint16_t id_;
     std::string name_;
-    size_t speed_;
-    size_t maxCrew_;
-    size_t capacity_;
-    size_t crew_;
+    uint16_t speed_;
+    uint16_t maxCrew_;
+    uint16_t capacity_;
+    uint16_t crew_;
 
 public:
     Ship();
-    Ship(size_t, const std::string&, size_t, size_t, size_t);
-    Ship(size_t, size_t, size_t);
+    Ship(uint16_t, const std::string&, uint16_t, uint16_t, uint16_t);
+    Ship(uint16_t, uint16_t, uint16_t);
 
     void setName(const std::string&);
 
-    Ship& operator+=(size_t);
-    Ship& operator-=(size_t);
+    Ship& operator+=(uint16_t);
+    Ship& operator-=(uint16_t);
 
-    size_t getId() const;
+    uint16_t getId() const;
     std::string getName();
-    size_t getSpeed() const;
-    size_t getMaxCrew() const;
-    size_t getCapacity() const;
+    uint16_t getSpeed() const;
+    uint16_t getMaxCrew() const;
+    uint16_t getCapacity() const;
 };
